Filename on the beginning of description
Subdivx does not have the filename as a separated field. This update searches on the description and put the filename on the beginning.
So "Los de aRGENTeaM para Iron.Man.2.720p.BluRay.x264-EXQUiSiTE" will be changed to "Iron.Man.2.720p.BluRay.x264-EXQUiSiTE Los de aRGENTeaM para " which makes easier to locate.